Add dinput8 and dxguid libs from DX8 into DX9 SDK.
Don't ask why, but I think the symbols for debug builds in the DX9 SDK
for these archives are corrupted. At linking the linker kept saying to
rebuild because of corrupted symbols. That didn't work, so I tried DX8
libs for that part and it worked. Hence I suffixed them with 'd' to
emphasize debug.
It could also be because we're still using old version of msvc6 using
DX9. So the linker can't interpret certain symbols. I'll revisit this
once we upgrade compilers.
I also git diff --no-index to see if there are difference between the
header files. They look the same, so it should be safe to use.